+++ src/constants/translateSource.js
@@ -1,9 +1,18 @@
-const translateSource = [
-    {source: 'google.com', url: 'translate.google.com'},
-    {source: 'google.cn', url: 'translate.google.cn'}
-];
-
 export const GOOGLE_CN = 'google.cn';
 export const GOOGLE_COM = 'google.com';
+export const BING_COM = 'bing.com';
+export const BING_CN = 'bing.cn';
+export const MOJIDICT_COM = 'mojidict.com';
 
-export default translateSource;
+export const translateSource = [
+    { source: GOOGLE_COM, url: 'translate.google.com' },
+    { source: GOOGLE_CN, url: 'translate.google.cn' },
+    { source: BING_COM, url: 'www.bing.com/translator/' },
+    { source: BING_CN, url: 'cn.bing.com/translator/' },
+    { source: MOJIDICT_COM, url: 'www.mojidict.com' }
+];
+
+export const audioSource = [
+    { source: GOOGLE_COM, url: 'translate.google.com' },
+    { source: GOOGLE_CN, url: 'translate.google.cn' }
+];