+++ src/entry/background/install.js
@@ -1,7 +1,6 @@
 import {LANG_ZH_CN, LANG_JA} from '../../constants/langCode';
 import {GOOGLE_CN} from '../../constants/translateSource';
 import defaultOptions from '../../constants/defaultOptions';
-import {initContextMenus} from './context-menus';
 import {getLocalStorage} from '../../public/chrome-call';
 
 /* global chrome */
@@ -14,6 +13,7 @@ const initStorageOnInstalled = (userLang) => {
         case LANG_ZH_CN:
             defaultSet.defaultTranslateSource = GOOGLE_CN;
             defaultSet.userLanguage = LANG_ZH_CN;
+            defaultSet.defaultAudioSource = GOOGLE_CN;
             break;
         case LANG_JA:
             defaultSet.userLanguage = LANG_JA;
@@ -30,6 +30,4 @@ chrome.runtime.onInstalled.addListener(() => {
     const userLang = navigator.language;
 
     initStorageOnInstalled(userLang);
-
-    initContextMenus(userLang);
 });