+++ src/entry/background/on-message.js
@@ -1,8 +1,15 @@
 import * as types from '../../constants/chromeSendMessageTypes';
 import {translate, audio} from '../../public/request';
 import {playAudio} from './audio';
+import { listenOptionsChange } from '../../public/options';
+import { getLocalStorage } from '../../public/chrome-call';
+import { GOOGLE_COM } from '../../constants/translateSource';
 
 /* global chrome */
+
+let defaultAudioSource = GOOGLE_COM;
+getLocalStorage('defaultAudioSource', storage => defaultAudioSource = storage.defaultAudioSource);
+listenOptionsChange(['defaultAudioSource'], changes => defaultAudioSource = changes.defaultAudioSource);
 
 chrome.runtime.onMessage.addListener(
     (request, sender, sendResponse) => {
@@ -17,6 +24,8 @@ chrome.runtime.onMessage.addListener(
                 return true;
             case types.SCTS_AUDIO:
                 if (payload) {
+                    !payload.source && (payload.source = defaultAudioSource);
+                    payload.defaultSource = defaultAudioSource;
                     audio(payload, uri => playAudio(uri));
                 }
                 return false;