+++ src/public/react-use/index.js
@@ -1,4 +1,3 @@
-export { default as useGetSelection} from './useGetSelection';
 export { default as useIsEnable } from './useIsEnable';
 export { default as useOnExtensionMessage } from './useOnExtensionMessage';
 export { default as useOptions } from './useOptions';