+++ CHANGELOG.md
@@ -1,5 +1,14 @@
+# 1.2.0
+- Add "translate" keyboard shortcut (default: Alt + Z).
+- Add option "show button after select".
+- Add "listen" keyboard shortcut (default: Alt + X).
+- Add translate source "bing.com".
+- Add translate source "mojidict.com".
+- Optimize get selection.
+- Fix "context menus invalid".
+- Separate "translate" and "listen".
 # 1.1.0
-- Support i18n (support en, ja, zh-cn, zh-tw now).
-- Support dark mode.
+- Add i18n (support en, ja, zh-cn, zh-tw now).
+- Add dark mode.
 # 1.0.0
 - Initial